[12.0][REF] website_sale_customer_type: externalize payment externalization made in `website_sale_customer_type_payment`
