+++ website_sale_customer_type/models/res_partner.py
@@ -25,15 +25,6 @@ class ResPartner(models.Model):
         string="Product",
         compute="_compute_website_product_ids",
     )
-    website_restrict_acquirer = fields.Boolean(
-        string="Restrict Acquirer on E-commerce",
-        compute="_compute_website_restrict_acquirer"
-    )
-    website_acquirer_ids = fields.Many2many(
-        comodel_name="payment.acquirer",
-        string="Acquirer",
-        compute="_compute_website_acquirer_ids",
-    )
 
     def get_customer_type_id(self):
         """
@@ -72,30 +63,3 @@ class ResPartner(models.Model):
             ):
                 products |= partner.get_customer_type_id().website_product_ids
             partner.website_product_ids = products
-
-    @api.depends("customer_type_id")
-    def _compute_website_restrict_acquirer(self):
-        """
-        Set website_restrict_acquirer to True if the customer type
-        has the flag website_restrict_acquirer set.
-        """
-        for partner in self:
-            partner.website_restrict_acquirer = (
-                partner.get_customer_type_id()
-                and partner.get_customer_type_id().website_restrict_acquirer
-            )
-
-    @api.depends("customer_type_id")
-    def _compute_website_acquirer_ids(self):
-        """
-        Return the list of acquirer that can be seen on the e-commerce by
-        the partner if connected depending on its customer type.
-        """
-        for partner in self:
-            acquirers = self.env["payment.acquirer"]
-            if (
-                partner.get_customer_type_id()
-                and partner.get_customer_type_id().website_acquirer_ids
-            ):
-                acquirers |= partner.get_customer_type_id().website_acquirer_ids
-            partner.website_acquirer_ids = acquirers
