+++ website_sale_customer_type_payment/__init__.py
@@ -1,0 +1,2 @@
+from . import models
+from . import controllers
