+++ website_sale_customer_type_payment/controllers/__init__.py
@@ -1,0 +1,1 @@
+from . import main
