+++ website_sale_customer_type_payment/models/__init__.py
@@ -1,4 +1,3 @@
 from . import res_partner_customer_type
 from . import res_partner
-from . import res_users
-from . import product_product
+from . import payment_acquirer
