+++ website_sale_customer_type_payment/readme/DESCRIPTION.rst
@@ -1,0 +1,2 @@
+This module restricts payment acquirers that a Customer Type can use on the e-commerce
+to a list of allowed acquirers.
