+++ website_sale_customer_type_payment/views/payment_views.xml
