+++ website_sale_customer_type_payment/views/res_partner_customer_type.xml
@@ -1,0 +1,24 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!-- Copyright 2021 Coop IT Easy SCRLfs
+     License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl). -->
+<odoo>
+
+    <record id="view_partner_customer_type_form" model="ir.ui.view">
+        <field name="name">res.partner.customer.type.payment.form</field>
+        <field name="model">res.partner.customer.type</field>
+        <field name="inherit_id" ref="website_sale_customer_type.view_partner_customer_type_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//notebook" position="inside">
+                <page string="Payment Acquirer">
+                    <group name="acquirer_config">
+                        <field name="website_restrict_acquirer"/>
+                        <field name="website_acquirer_ids"
+                               attrs="{'invisible': [('website_restrict_acquirer', '=', False)]}"
+                               widget="many2many_tags"/>
+                    </group>
+                </page>
+            </xpath>
+        </field>
+    </record>
+
+</odoo>
